Updated SQL, new SQL and documents
Renamed the SQL to create the hourly_airtemp view.  Added in a document
which tells more about units in ‘DataValues’  table.  Added in updated
SQL for the seriesCatalog_62.  New SQL to update the ‘Datastreams’
table, create the daily air temp view, insert daily air temp and the
usp for the daily air temp.